Create PXE_container_deploy_template.yml
Created initial template. For now variables should be assigned with an independant inventory for each CT instance. This will be used for role creation in the future.

Notes: This hasn't been ran yet, there could be an issue with the netif statement, I'm unsure at the moment if variables can nested in other variables.